added 2 new select such as: user_id and owner_id for QueueListView class. And checks new permission

+++ backend/service/Queue/views.py
@@ -17,6 +17,7 @@ from rest_framework.response import Response
 from rest_framework.settings import api_settings
 from rest_framework.viewsets import ModelViewSet
 
+from JWTUser.permissions import IsOwnerQueue
 from Queue.models import Queue, UserQueueRelation
 from Queue.serializers import QueueSerializer, UserQueueRelationRegisterSerializer
 
@@ -57,32 +58,53 @@ class QueueRegisterView(CreateAPIView):
 class QueueListView(ListAPIView):
     queryset = Queue.objects.all()
     serializer_class = QueueSerializer
-    permission_classes = (AllowAny,)
+    permission_classes = (IsAuthenticated,)
+
+    def get_queryset(self):
+
+        queryset = Queue.objects.all()
+        user_id = self.request.query_params.get('user_id', None)
+        owner_id = self.request.query_params.get('owner_id', None)
+
+        if user_id is not None and owner_id is not None:
+            queryset = (queryset.filter(users__id=user_id) | queryset.filter(owner__id=owner_id)).distinct()
+        elif user_id is not None:
+            queryset = queryset.filter(users__id=user_id)
+        elif owner_id is not None:
+            queryset = queryset.filter(owner_id=owner_id)
+
+        return queryset
 
 
 class QueueRetrieveView(RetrieveAPIView):
     queryset = Queue.objects.all()
     serializer_class = QueueSerializer
-    permission_classes = (AllowAny,)
+    permission_classes = (IsAuthenticated,)
 
 
 class UserQueueRelationRegisterView(CreateAPIView):
     queryset = UserQueueRelation.objects.all()
     serializer_class = UserQueueRelationRegisterSerializer
-    permission_classes = (AllowAny,)
+    permission_classes = (IsAuthenticated,)
 
 
 class UserQueueRelationDestroyView(DestroyAPIView):
     queryset = UserQueueRelation.objects.all()
     serializer_class = UserQueueRelationRegisterSerializer
-    permission_classes = (AllowAny,)
+    permission_classes = (IsOwnerQueue, IsAuthenticated)
 
     def delete(self, request, *args, **kwargs):
+        self.check_permissions(request)
         user_id = kwargs.get('user_id')
         queue_id = kwargs.get('queue_id')
         try:
+            queue = Queue.objects.get(pk=queue_id)
+            self.check_object_permissions(
+                request=self.request,
+                obj=queue,
+            )
             user_queue_relation = UserQueueRelation.objects.filter(user=user_id, queue=queue_id).last()
             user_queue_relation.delete()
             return Response(status=status.HTTP_204_NO_CONTENT)
-        except (UserQueueRelation.DoesNotExist, AttributeError):
+        except (UserQueueRelation.DoesNotExist, Queue.DoesNotExist, AttributeError):
             return Response({'detail': 'UserQueueRelation not found'}, status=status.HTTP_404_NOT_FOUND)
